refactor: split schema into modules
