+++ crates/tenx-mcp/src/schema/mod.rs
@@ -1,0 +1,7 @@
+pub mod jsonrpc;
+pub mod types;
+pub mod requests;
+
+pub use jsonrpc::*;
+pub use types::*;
+pub use requests::*;
